OpenSourceModels/GEMINI IMAGE TO TEXT/requirements.txt
streamlit
google_generativeai
pillow